requirements.txt
tensorflow
opencv-python
numpy
matplotlib
streamlit
pillow
scikit-learn
